Move arguments to their own file
