+++ crates/orderbook/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod account_balances;
 pub mod api;
+pub mod arguments;
 pub mod conversions;
 pub mod cow_subsidy;
 pub mod database;
